Remove compiler warning.

+++ min/humlib.cpp
@@ -1,7 +1,7 @@
 //
 // Programmer:    Craig Stuart Sapp <craig@ccrma.stanford.edu>
 // Creation Date: Sat Aug  8 12:24:49 PDT 2015
-// Last Modified: Tue May 14 03:29:40 PDT 2024
+// Last Modified: Tue May 14 04:19:39 PDT 2024
 // Filename:      min/humlib.cpp
 // URL:           https://github.com/craigsapp/humlib/blob/master/min/humlib.cpp
 // Syntax:        C++11
@@ -117402,7 +117402,7 @@ vector<string> Tool_shed::addToExInterpList(void) {
 	hre.split(pieces, elist, "[,;\\s*]+");
 
 	vector<string> output;
-	for (int i=0; i<pieces.size(); i++) {
+	for (int i=0; i<(int)pieces.size(); i++) {
 		if (pieces[i].empty()) {
 			continue;
 		}
